+++ src/lib.rs
@@ -3,6 +3,7 @@
 #![cfg_attr(feature = "nightly", feature(duration_float))]
 
 mod boxed;
+mod hash;
 mod collections;
 mod comprehension;
 mod times;